:recycle: refactor: 파일 디렉터리 구조 및 위치 정리
